Support "Old" string concat behavior
Prior to v5.1, NCalc would treat + as string concatenation if the first argument was a string. Later versions changed this logic and while they did add options to control it, no combination of those options restored the old behavior.

The Tech Software fork exists to allow callers to opt-in to the old behavior.